Refactor runners to use ProcessRunner helper
